move more series tests to test_arithmetic
