+++ pandas/tests/series/test_timezones.py
@@ -88,6 +88,15 @@ class TestSeriesTimezones(object):
         tm.assert_raises_regex(TypeError, "Cannot convert tz-naive",
                                ts.tz_convert, 'US/Eastern')
 
+    def test_series_tz_convert_to_utc(self):
+        base = DatetimeIndex(['2011-01-01', '2011-01-02', '2011-01-03'],
+                             tz='UTC')
+        idx1 = base.tz_convert('Asia/Tokyo')[:2]
+        idx2 = base.tz_convert('US/Eastern')[1:]
+
+        res = Series([1, 2], index=idx1) + Series([1, 1], index=idx2)
+        tm.assert_series_equal(res, Series([np.nan, 3, np.nan], index=base))
+
     # -----------------------------------------------------------------
     # Series.append
